Split readme: one for GitHub/PyPI, one for docs
The readme file for GitHub/PyPI is README.md. It contains a short summary, a
quickstart and an author section. markdown is enough.

The docs/index.rst file contains the contents of the original README.rst file.
This makes the first page of the documentation much nicer and we don't have a
separate readme page anymore.