Fix datalim problem when rendering
BallCollection.get_datalim was not working correctly because of the it handles
coordinate transformations. I suspect it can still be improved (I don't quite
understand how transform, _transforms and transOffset are connected).
I have included BallPatchCollection for reference, this one will correctly
render but uses a new path for every ball in the scene (i.e. it is slower and
uses more memory than necessary).