Use different initial times for each ball
In a previous commit we fixed some floating point issues by introducing
`_last_time` and `_last_balls_position` to record the position of all balls at
the last simulation-wide collision. In this commit we basically split up
`_last_time` across each ball so that the initial ball positions are not reset
each time there is a collision somewhere in the simulation. We still recompute
the positions at each collision, but since numpy takes care of this has almost
no effect on speed (~5% slower).

The properties have been renamed to `balls_initial_time` and
`balls_initial_position` and `balls_position` is computed from them.

Also in `recompute_toi` we check which `balls_position` were modified and reset
the initial properties accordingly.