Rework development workflow
- Remove pre-commit from requirements_dev.txt and Pipfile, it is useful but
optional for development.
- Move flake8 configuration from tox.ini to its own file (because it is used in
pre-commit and not tox). Also, remove the flake8 format setting and the
flake8-colors plugin, because the default format is good enough and doesn't
become unreadable when flake8-colors is not installed.
- MANIFEST.in: Exclude example pictures and videos from source distribution.
Also we do not need to exclude docs here, instead I configure the ignore
setting for check-manifest in tox.ini to ignore the first two levels of the
docs folders (check-manifest cannot ignore folders recursively!?!)
- When generating the documentation, I get a warning about linkcheck behaviour
in Sphinx>=8.0. I added a setting to docs/conf.py just to be sure.
- Clarify tox environments for linting and building the documentation. Add
pydocstyle and doc8 in the linting step (they are also used in pre-commit).
- Rewrite CONTRIBUTING.md, adding the comments above and changing indentation
in some sections.