Improve floating point accuracy for Circle
- Move the circle collision detection to physics.py
- Compute upper bounds for the rounding error of dist_sqrd
- Check first where the ball is (accounting for rounding errors), then call
toi_ball_ball or toi_ball_disk_exterior to compute the time of impact
- Add more tests, especially for point particles

Caveats:
- A ball that is touching the circle from the inside and moving towards the
outside will not collide with the circle. This situation shouldn't occur in
practice because we reflect the velocity vector after collisions.
- The upper bounds on the rounding error assume that all vectors contain
double-precision floating point numbers and that they are not affected by
previous rounding errors. The last assumption may not hold in practice due to
cancellation of significant digits, e.g. when computing a collision at the new
position pos + t * vel, but pos, vel and t are large and such that pos + t *
vel ends up close to the origin.